requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorlog==6.7.0
cssselect==1.2.0
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
fake-useragent==1.2.1
flake8==6.1.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
mccabe==0.7.0
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parse==1.19.1
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pyee==8.2.2
pyflakes==3.1.0
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-html==0.10.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
tzdata==2023.3
urllib3==1.26.16
w3lib==2.1.2
webdriver-manager==4.0.0
websockets==10.4
wsproto==1.2.0
zipp==3.16.2
